refactor(ssh): split gateway adapter responsibilities
